Create ICanifier and impl, move classes to appropriate packages
