+++ src/main/java/com/team1816/lib/hardware/CtreMotorFactory.java
@@ -4,8 +4,8 @@ import com.ctre.phoenix.ErrorCode;
 import com.ctre.phoenix.ParamEnum;
 import com.ctre.phoenix.motorcontrol.*;
 import com.ctre.phoenix.motorcontrol.can.*;
-import com.team1816.lib.hardware.components.GhostTalonSRX;
-import com.team1816.lib.hardware.components.LazyTalonFX;
+import com.team1816.lib.hardware.components.motor.GhostTalonSRX;
+import com.team1816.lib.hardware.components.motor.LazyTalonFX;
 import com.team254.lib.drivers.LazyTalonSRX;
 
 /**
