--- src/main/java/com/team1816/lib/hardware/components/GhostCanifier.java
@@ -1,4 +1,0 @@
-package com.team1816.lib.hardware.components;
-
-public interface GhostCanifier {
-}
