+++ src/main/java/com/team1816/lib/hardware/components/motor/GhostTalonSRX.java
@@ -1,4 +1,4 @@
-package com.team1816.lib.hardware.components;
+package com.team1816.lib.hardware.components.motor;
 
 import com.ctre.phoenix.ErrorCode;
 import com.ctre.phoenix.ParamEnum;
