+++ src/main/java/com/team1816/lib/hardware/components/motor/ILazyMotorControllerEnhanced.java
@@ -1,4 +1,4 @@
-package com.team1816.lib.hardware.components;
+package com.team1816.lib.hardware.components.motor;
 
 import com.ctre.phoenix.motorcontrol.IMotorControllerEnhanced;
