+++ src/main/java/com/team1816/lib/hardware/components/motor/LazyTalonFX.java
@@ -1,4 +1,4 @@
-package com.team1816.lib.hardware.components;
+package com.team1816.lib.hardware.components.motor;
 
 import com.ctre.phoenix.motorcontrol.ControlMode;
 import com.ctre.phoenix.motorcontrol.can.TalonFX;
