+++ src/main/java/com/team1816/lib/hardware/components/pcm/DoubleSolenoidImpl.java
@@ -1,0 +1,4 @@
+package com.team1816.lib.hardware.components.pcm;
+
+public class DoubleSolenoidImpl {
+}
