+++ src/main/java/com/team1816/lib/hardware/components/pcm/ISolenoid.java
@@ -1,0 +1,5 @@
+package com.team1816.lib.hardware.components.pcm;
+
+public interface ISolenoid {
+
+}
